+++ eyeball/src/read_guard.rs
@@ -1,8 +1,6 @@
 use std::{fmt, ops};
 
-use readlock::SharedReadGuard;
-
-use crate::state::ObservableState;
+use crate::{lock::Lock, state::ObservableState, SyncLock};
 
 /// A read guard for the inner value of an observable.
 ///
@@ -10,23 +8,23 @@ use crate::state::ObservableState;
 /// `Observable` is locked and can not be updated.
 #[must_use]
 #[clippy::has_significant_drop]
-pub struct ObservableReadGuard<'a, T> {
-    inner: SharedReadGuard<'a, ObservableState<T>>,
+pub struct ObservableReadGuard<'a, T: 'a, L: Lock = SyncLock> {
+    inner: L::SharedReadGuard<'a, ObservableState<T>>,
 }
 
-impl<'a, T> ObservableReadGuard<'a, T> {
-    pub(crate) fn new(inner: SharedReadGuard<'a, ObservableState<T>>) -> Self {
+impl<'a, T: 'a, L: Lock> ObservableReadGuard<'a, T, L> {
+    pub(crate) fn new(inner: L::SharedReadGuard<'a, ObservableState<T>>) -> Self {
         Self { inner }
     }
 }
 
-impl<T: fmt::Debug> fmt::Debug for ObservableReadGuard<'_, T> {
+impl<T: fmt::Debug, L: Lock> fmt::Debug for ObservableReadGuard<'_, T, L> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.inner.fmt(f)
     }
 }
 
-impl<T> ops::Deref for ObservableReadGuard<'_, T> {
+impl<T, L: Lock> ops::Deref for ObservableReadGuard<'_, T, L> {
     type Target = T;
 
     fn deref(&self) -> &Self::Target {